backend/adoorback/requirements.txt
amqp==5.0.2
arrow==0.17.0
asgiref==3.2.10
astroid==2.4.2
autopep8==1.5.4
awesome-slugify==1.6.5
beautifulsoup4==4.6.0
billiard==3.6.3.0
binaryornot==0.4.4
certifi==2020.6.20
cffi==1.14.4
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
colorama==0.4.4
ConfigArgParse==1.2.3
cookiecutter==1.7.2
cookiecutter-django==1.11.9
coverage==5.3
coveralls==2.1.2
cryptography==3.3.1
defusedxml==0.6.0
diff-match-patch==20200713
dill==0.3.3
Django==3.1.2
django-admin-honeypot==1.1.0
django-common-helpers==0.9.2
django-cors-headers==3.5.0
django-cron==0.5.1
django-extensions==3.0.9
django-import-export==2.4.0
django-polymorphic==3.0.0
django-queryset-csv==1.1.0
django-redis==4.12.1
django-rest-polymorphic==0.1.9
django-test-plus==1.4.0
django-trackstats==0.5.0
djangorestframework==3.12.1
djangorestframework-simplejwt==4.6.0
docopt==0.6.2
environ==1.0
et-xmlfile==1.0.1
Faker==4.14.0
Flask==1.1.2
Flask-BasicAuth==0.2.0
gevent==20.9.0
geventhttpclient==1.4.4
greenlet==0.4.17
idna==2.10
importlib-metadata==2.0.0
isort==5.6.4
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.11.2
jinja2-time==0.2.0
joblib==0.17.0
JPype1==1.1.2
kombu==5.0.2
konlpy==0.5.2
lazy-object-proxy==1.4.3
locust==1.4.1
lxml==4.6.1
MarkupPy==1.14
MarkupSafe==1.1.1
mccabe==0.6.1
mecab-python3==1.0.3
msgpack==1.0.0
ndg-httpsclient==0.5.1
nltk==3.5
numpy==1.19.4
oauthlib==3.1.0
odfpy==1.4.1
openpyxl==3.0.5
pandas==1.1.4
pathlib2==2.3.5
Pillow==8.0.1
poyo==0.5.0
prompt-toolkit==3.0.8
psutil==5.7.3
pyasn1==0.4.8
pycodestyle==2.6.0
pycparser==2.20
Pygments==2.7.2
PyJWT==1.7.1
pylev==1.3.0
pylint==2.6.0
pylint-django==2.3.0
pylint-plugin-utils==0.6
pyOpenSSL==20.0.1
PySocks==1.7.1
python-dateutil==2.8.1
python-slugify==4.0.1
pytz==2020.1
PyYAML==5.3.1
pyzmq==20.0.0
redis==3.5.3
regex==2020.10.28
requests==2.24.0
requests-oauthlib==1.3.0
scikit-learn==0.23.2
scipy==1.5.4
sentry-sdk==0.19.4
six==1.15.0
sqlparse==0.4.1
tablib==2.0.0
text-unidecode==1.3
threadpoolctl==2.1.0
toml==0.10.1
tqdm==4.53.0
tweepy==3.9.0
typed-ast==1.4.1
typing-extensions==3.7.4.3
unicodecsv==0.14.1
Unidecode==0.4.21
urllib3==1.25.11
vine==5.0.0
wcwidth==0.2.5
Werkzeug==1.0.1
wrapt==1.12.1
xlrd==1.2.0
xlwt==1.3.0
zipp==3.4.0
zope.event==4.5.0
zope.interface==5.2.0
